requirements.txt
absl-py==2.3.1
accelerate==1.9.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
autocommand==2.2.2
av==15.0.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
blinker==1.9.0
blis==1.2.1
botocore==1.40.25
braceexpand==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2022.12.7
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
confection==0.1.5
ConfigArgParse==1.7
contexttimer==0.3.3
contourpy==1.3.0
cupy-cuda12x==13.3.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
debugpy==1.8.11
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
distlib==0.4.0
docker-pycreds==0.4.0
docopt==0.6.2
dominate==2.9.1
durationpy==0.10
easydict==1.13
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.116.1
fastjsonschema==2.21.1
fastrlock==0.8.2
ffmpeg==1.4
ffmpy==0.6.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.59.2
freetype-py==2.5.1
frozenlist==1.7.0
fschat==0.2.36
fsspec==2024.2.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.40.3
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
hydra-core==1.3.2
identify==2.6.12
idna==3.4
imageio==2.35.1
imageio-ffmpeg==0.6.0
importlib_metadata==8.0.0
importlib_resources==6.5.2
inflect==7.3.1
inotify-simple==2.0.1
iopath==0.1.9
ipykernel==6.29.5
ipython==8.12.3
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaggle==1.7.4.5
keras==3.10.0
kiwisolver==1.4.7
kubernetes==33.1.0
langcodes==3.5.0
language_data==1.3.0
latex2mathml==3.78.0
lazy_loader==0.4
libclang==18.1.1
linkify-it-py==2.0.3
llvmlite==0.43.0
loguru==0.7.2
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memory-profiler==0.61.0
memray==1.19.1
mistune==3.1.4
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.3.0
ninja==1.11.1.1
nodeenv==1.9.1
ntplib==0.4.0
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python==4.10.0.82
opencv-python-headless==4.11.0.86
opendatasets==0.1.22
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.1
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.2
plotly==6.2.0
ply==3.11
portalocker==2.10.1
pre_commit==4.2.0
preshed==3.0.10
proglog==0.1.12
prompt_toolkit==3.0.48
propcache==0.3.2
protobuf==5.28.3
psutil==6.0.0
ptflops==0.7.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyglet==2.0.17
Pygments==2.19.2
PyOpenGL==3.1.0
pyparsing==3.2.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.7
s3transfer==0.13.1
safetensors==0.5.3
scalene==1.5.55
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.18.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
simplejson==3.20.1
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.2
streamlit==1.47.1
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.20.0
termcolor==2.5.0
text-unidecode==1.3
textual==6.5.0
thinc==8.3.4
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.9.0
timm==1.0.19
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
torch==2.5.0+cu121
torch-lr-finder==0.2.2
torchaudio==2.5.0+cu121
torchvision==0.20.0+cu121
tornado==6.5.2
tqdm==4.66.5
traitlets==5.14.3
trimesh==4.4.9
triton==3.1.0
typeguard==4.3.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==1.26.20
uvicorn==0.35.0
virtualenv==20.32.0
visdom==0.2.4
vit-pytorch==1.16.1
wandb==0.18.6
wasabi==1.1.3
watchdog==6.0.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
yacs==0.1.8
yarg==0.1.9
yarl==1.20.1
zipp==3.23.0
